Сделал задания: добавил модули, добавил скрипт в index, для наглядности работоспособности импортировал arrayOffers и вывел в консоль
